Remove AWS based autoscaler code
We still support metrics based autoscaling.

Fixes https://github.com/cortexproject/cortex/issues/1379
